backend/requirements.txt
asgiref==3.3.4
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-restql==0.11.2
django-url-filter==0.3.15
djangorestframework==3.12.4
djangorestframework-queryfields==1.0.0
drf-writable-nested==0.6.2
enum-compat==0.0.3
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
pyPEG2==2.15.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
